Refactored into a single component where the export is the container
